+++ freeton/contracts/interfaces/IFabric.sol
@@ -6,9 +6,9 @@ interface IFabric {
     function onPoolDeploy(
         uint64 pool_deploy_nonce,
         address pool_owner,
-        uint256[] rewardPerSecond,
-        uint256 farmStartTime,
-        uint256 farmEndTime,
+        uint128[] rewardPerSecond,
+        uint32 farmStartTime,
+        uint32 farmEndTime,
         address tokenRoot,
         address[] rewardTokenRoot
     ) external;
